Update `isReadyForAnalysis` logic for C++.
For C/C++ targets, we don't track or explicitly build external deps. Instead, we need to gather cc compilation info for each in-project target. To support this in the dependnecy tracker, introduce a concept of `DependencyTrackingBehavior` which helps keep the other logic language agnostic. This is a (small) step towards making support for more languages easier in future.

Update the `getPendingTargets` logic based on the dependency tracking behaviour, only gathering transitive external dependnecies when required.

This ensures that we enable analysis for cc targets whenever we have build deps for the target itself.

PiperOrigin-RevId: 583338131